+++ 2O2P/src/oo2cad/cad/objects/CadBow.java
@@ -7,12 +7,12 @@ package oo2cad.cad.objects;
  */
 public class CadBow extends CadBaseObject
 {
-	private float radiusX;
-	private float radiusY;
-	private float angleStart;
-	private float angleEnd;
+	private double radiusX;
+	private double radiusY;
+	private double angleStart;
+	private double angleEnd;
 	
-	public CadBow(float radiusX, float radiusY, float angleStart, float angleEnd)
+	public CadBow(double radiusX, double radiusY, double angleStart, double angleEnd)
 	{
 		this.radiusX = radiusX;
 		this.radiusY = radiusY;
@@ -20,48 +20,48 @@ public class CadBow extends CadBaseObject
 		this.angleEnd = angleEnd;
 	}
 	
-	public CadBow(float radiusX, float radiusY)
+	public CadBow(double radiusX, double radiusY)
 	{
 		this.radiusX = radiusX;
 		this.radiusY = radiusY;
 	}
 
-	public float getRadiusX()
+	public double getRadiusX()
 	{
 		return radiusX;
 	}
 
-	public void setRadiusX(float radiusX)
+	public void setRadiusX(double radiusX)
 	{
 		this.radiusX = radiusX;
 	}
 
-	public float getRadiusY()
+	public double getRadiusY()
 	{
 		return radiusY;
 	}
 
-	public void setRadiusY(float radiusY)
+	public void setRadiusY(double radiusY)
 	{
 		this.radiusY = radiusY;
 	}
 
-	public float getAngleStart()
+	public double getAngleStart()
 	{
 		return angleStart;
 	}
 
-	public void setAngleStart(float angleStart)
+	public void setAngleStart(double angleStart)
 	{
 		this.angleStart = angleStart;
 	}
 
-	public float getAngleEnd()
+	public double getAngleEnd()
 	{
 		return angleEnd;
 	}
 
-	public void setAngleEnd(float angleEnd)
+	public void setAngleEnd(double angleEnd)
 	{
 		this.angleEnd = angleEnd;
 	}
