+++ 2O2P/src/oo2cad/shapes/SimpleShape.java
@@ -2,40 +2,40 @@ package oo2cad.shapes;
 
 public class SimpleShape extends Shape
 {
-	private float startX;
-	private float startY;
-	private float endX;
-	private float endY;
+	private double startX;
+	private double startY;
+	private double endX;
+	private double endY;
 	
-	public float getStartX()
+	public double getStartX()
 	{
 		return startX;
 	}
-	public void setStartX(float startX)
+	public void setStartX(double startX)
 	{
 		this.startX = startX;
 	}
-	public float getStartY()
+	public double getStartY()
 	{
 		return startY;
 	}
-	public void setStartY(float startY)
+	public void setStartY(double startY)
 	{
 		this.startY = startY;
 	}
-	public float getEndX()
+	public double getEndX()
 	{
 		return endX;
 	}
-	public void setEndX(float endX)
+	public void setEndX(double endX)
 	{
 		this.endX = endX;
 	}
-	public float getEndY()
+	public double getEndY()
 	{
 		return endY;
 	}
-	public void setEndY(float endY)
+	public void setEndY(double endY)
 	{
 		this.endY = endY;
 	}
